users, login, registration UI + STABLE BUILD
this is a redo of all this code because we lost a bunch of it on the old server. lots of little bug fixes and playing around with.

the clock has been fixed up.

the last version actually had a jpg bug so the binary wouldnt boot.

syslog.txt for all your boot errors.

lots of work has been spent on multi-threading support. i will start a new branch for this work. we really need multi threading support in aura so we can take it further for tcp applications.